Rearrange Jackson Datatype Folders
