--- jackson-datatypes/jackson-datatype-interledger-address/src/main/java/org/interledger/quilt/jackson/InterledgerAddressModule.java
@@ -1,30 +1,0 @@
-package org.interledger.quilt.jackson;
-
-import org.interledger.core.InterledgerAddress;
-
-import com.fasterxml.jackson.core.Version;
-import com.fasterxml.jackson.databind.module.SimpleModule;
-
-/**
- * A Jackson {@link SimpleModule} for serializing and deserializing {@link InterledgerAddress}
- * objects.
- */
-public class InterledgerAddressModule extends SimpleModule {
-
-  private static final String NAME = "InterledgerAddressModule";
-
-  /**
-   * No-args Constructor.
-   */
-  public InterledgerAddressModule() {
-
-    super(
-        NAME,
-        new Version(1, 0, 0, null, "org.interledger", "jackson-datatype-interledger-address")
-    );
-
-    addSerializer(InterledgerAddress.class, InterledgerAddressSerializer.INSTANCE);
-    addDeserializer(InterledgerAddress.class, InterledgerAddressDeserializer.INSTANCE);
-  }
-
-}
