--- jackson-datatypes/jackson-datatype-interledger/pom.xml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project xmlns="http://maven.apache.org/POM/4.0.0"
-  xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
-  xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <parent>
-    <artifactId>jackson-datatypes</artifactId>
-    <groupId>org.interledger</groupId>
-    <version>0.2.0-SNAPSHOT</version>
-  </parent>
-  <modelVersion>4.0.0</modelVersion>
-
-  <name>Jackson Interledger Module</name>
-  <artifactId>jackson-datatype-interledger</artifactId>
-
-  <dependencies>
-    <dependency>
-      <groupId>org.interledger</groupId>
-      <artifactId>ilp-core</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>org.interledger</groupId>
-      <artifactId>jackson-datatype-interledger-address</artifactId>
-    </dependency>
-    <dependency>
-      <groupId>com.google.guava</groupId>
-      <artifactId>guava</artifactId>
-      <!-- Marking as provided because this dependency is not needed unless Base16 Hex Encoding is used. -->
-      <scope>provided</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.bouncycastle</groupId>
-      <artifactId>bcprov-jdk16</artifactId>
-      <scope>test</scope>
-    </dependency>
-    <dependency>
-      <groupId>org.immutables</groupId>
-      <artifactId>value</artifactId>
-      <scope>provided</scope>
-    </dependency>
-  </dependencies>
-
-</project>
