+++ jackson-datatypes/src/main/java/org/interledger/quilt/jackson/conditions/ConditionSerializer.java
